Refactor: adapter들 ui 폴더 내로 이동
