+++ app/src/main/java/com/depromeet/housekeeper/ui/houseRule/adapter/RuleAdapter.kt
@@ -1,4 +1,4 @@
-package com.depromeet.housekeeper.adapter
+package com.depromeet.housekeeper.ui.houseRule.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
