+++ app/src/main/java/com/depromeet/housekeeper/ui/main/adapter/SmallProfileAdapter.kt
@@ -1,4 +1,4 @@
-package com.depromeet.housekeeper.adapter
+package com.depromeet.housekeeper.ui.main.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
